requirements.txt
psycopg2==2.8.6
psycopg2-binary==2.8.6
Django==3.1.3
requests==2.25.1
djangorestframework==3.12.1
mysqlclient==2.0.1
openpyxl==3.0.5
gunicorn==20.0.4
django-mathfilters==1.0.0
django-filter==2.4.0
python-dateutil==2.8.1
mock==4.0.2
coverage==5.4